M2.1c hardening: i32 coefficient pipeline + Huffman short-buffer path
Two Debug-mode panics surfaced by the comptime-Debug fail() instrumentation
on the real-world JPEG corpus (4,125 files at /Volumes/Fileserver/clips-image/):

1. i16 coefficient overflow: prev_dc and dequantized coeffs now flow as i32
   end-to-end (idct8x8 input widened, dequant cast removed). DC*qt can reach
   ~261k on real JPEGs which exceeds i16's ±32767 range. ReleaseFast wrapped
   silently; Debug aborted with integer-overflow trap.

2. Huffman fast-path consumed phantom bits: peekBits(8) zero-pads when fewer
   than 8 valid bits remain. fast_size[peek] could request more bits than
   bits_valid, tripping the consume() assert. Added a short-buffer path that
   walks bit-by-bit via readBits(1), surfacing NotEnoughBits cleanly instead
   of asserting.

Net corpus impact: Debug runs no longer crash (panics now graceful errors).
ReleaseFast CLEAN-OK count unchanged at 432/4125 (10.5%) — confirms these
were safety hardening, not coverage gains. Dominant remaining failure is
ac_huffman_decode_failed at 2249/2257 = 99.6% of CLEAN-ERR; that's the next
real bug to find.

+++ src/decode/idct.zig
@@ -47,9 +47,11 @@ const NORM: [N]f64 = blk: {
 /// Run the inverse 8×8 DCT on `coeffs` and write the spatial samples
 /// into `out`. Output is level-shifted by +128 and clamped to 0..255.
 ///
-/// `coeffs[v * 8 + u]` = the (u, v) frequency coefficient.
+/// `coeffs[v * 8 + u]` = the (u, v) frequency coefficient (post-dequant
+/// in natural row-major order; type i32 so adversarial DC*qt products
+/// don't overflow i16's ±32767 range).
 /// `out[y * 8 + x]` = the (x, y) spatial sample.
-pub fn idct8x8(coeffs: *const [64]i16, out: *[64]u8) void {
+pub fn idct8x8(coeffs: *const [64]i32, out: *[64]u8) void {
     var y: usize = 0;
     while (y < N) : (y += 1) {
         var x: usize = 0;
@@ -80,14 +82,14 @@ pub fn idct8x8(coeffs: *const [64]i16, out: *[64]u8) void {
 // ── Tests ────────────────────────────────────────────────────
 
 test "idct8x8 of all-zero coefficients yields uniform 128" {
-    const coeffs = [_]i16{0} ** 64;
+    const coeffs = [_]i32{0} ** 64;
     var out: [64]u8 = undefined;
     idct8x8(&coeffs, &out);
     for (out) |s| try std.testing.expectEqual(@as(u8, 128), s);
 }
 
 test "idct8x8 of DC-only with positive amplitude yields uniform brighter" {
-    var coeffs = [_]i16{0} ** 64;
+    var coeffs = [_]i32{0} ** 64;
     coeffs[0] = 256; // DC coefficient: shifts the mean up
     var out: [64]u8 = undefined;
     idct8x8(&coeffs, &out);
@@ -97,7 +99,7 @@ test "idct8x8 of DC-only with positive amplitude yields uniform brighter" {
 }
 
 test "idct8x8 of DC-only with negative amplitude yields uniform darker" {
-    var coeffs = [_]i16{0} ** 64;
+    var coeffs = [_]i32{0} ** 64;
     coeffs[0] = -256;
     var out: [64]u8 = undefined;
     idct8x8(&coeffs, &out);
@@ -106,7 +108,7 @@ test "idct8x8 of DC-only with negative amplitude yields uniform darker" {
 }
 
 test "idct8x8 clamps to 0..255" {
-    var coeffs = [_]i16{0} ** 64;
+    var coeffs = [_]i32{0} ** 64;
     coeffs[0] = 4096; // way too large; output would overflow without clamp
     var out: [64]u8 = undefined;
     idct8x8(&coeffs, &out);
